% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexgsea.R
\name{flexgsea_calc_sig_simple}
\alias{flexgsea_calc_sig_simple}
\title{Calculate significance of gene set enrichments from permutations.}
\usage{
flexgsea_calc_sig_simple
}
\description{
Calculates significance by the rank of ES score in the permuted values.
Significance is calculated separately per gene set.
}
\seealso{
Other functions for significance calculation: \code{\link{flexgsea_calc_sig}}
}
\concept{functions for significance calculation}
