% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexgsea_limma.R
\docType{data}
\name{flexgsea_limma_trend}
\alias{flexgsea_limma_trend}
\title{Score genes with limma-trend}
\format{
An object of class \code{list} of length 2.
}
\usage{
flexgsea_limma_trend
}
\description{
Scores genes by their coefficients in from limma. The data x given to flexgsea should be a matrix
with RNAseq counts. This function uses TMM normalization converts the data to counts per million
before running limma::lmFit. The response variables  y should be a design matrix, as produced by
\code{model.matrix}.
}
\seealso{
Other gene scoring functions: 
\code{\link{flexgsea_limma_voom}},
\code{\link{flexgsea_limma}},
\code{\link{flexgsea_lm}()},
\code{\link{flexgsea_s2n}()}
}
\concept{gene scoring functions}
\keyword{datasets}
