% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexgsea.R
\name{flexgsea_lm}
\alias{flexgsea_lm}
\title{Score genes in a linear regression model.}
\usage{
flexgsea_lm
}
\description{
Scores genes by their coefficients in a linear regression
model of \option{y} onto \option{x}. When there is one response variable
this is equivalent to Pearson correlation.
Do not call directly, but give as the \option{gene.score.fn}  argument to
\code{\link{flexgsea}}.
}
\seealso{
Other gene scoring functions: 
\code{\link{flexgsea_limma_trend}},
\code{\link{flexgsea_limma_voom}},
\code{\link{flexgsea_limma}},
\code{\link{flexgsea_s2n}()}
}
\concept{gene scoring functions}
