% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexgsea.R
\docType{data}
\name{flexgsea_maxmean}
\alias{flexgsea_maxmean}
\title{The maxmean statistic for calculating gene set enrichement scores.}
\format{An object of class \code{list} of length 3.}
\usage{
flexgsea_maxmean
}
\description{
The maxmean statistic for calculating gene set enrichement scores.
}
\seealso{
Other gene set enrichment functions: \code{\link{flexgsea_mean}},
  \code{\link{flexgsea_weighted_ks}}
}
\concept{gene set enrichment functions}
\keyword{datasets}
