% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexgsea.R
\docType{data}
\name{flexgsea_mean}
\alias{flexgsea_mean}
\title{The mean statistic for calculating gene set enrichement scores.}
\format{
An object of class \code{list} of length 3.
}
\usage{
flexgsea_mean
}
\description{
The mean statistic for calculating gene set enrichement scores.
}
\seealso{
Other gene set enrichment functions: 
\code{\link{flexgsea_maxmean}},
\code{\link{flexgsea_weighted_ks}}
}
\concept{gene set enrichment functions}
\keyword{datasets}
