% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexgsea.R
\name{flexgsea_s2n}
\alias{flexgsea_s2n}
\title{Score genes by there signal to noise ratio.}
\usage{
flexgsea_s2n
}
\description{
Scores genes by the signal to noise ration (s2n) a. Requires \option{y}
to have two classes.
Do not call directly, but give as the \option{gene.score.fn}  argument to
\code{\link{flexgsea}}.
}
\seealso{
Other gene scoring functions: 
\code{\link{flexgsea_limma_trend}},
\code{\link{flexgsea_limma_voom}},
\code{\link{flexgsea_limma}},
\code{\link{flexgsea_lm}()}
}
\concept{gene scoring functions}
