% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexgsea.R
\docType{data}
\name{flexgsea_weighted_ks}
\alias{flexgsea_weighted_ks}
\title{The weighted KS statistic for calculating gene set enrichement scores.}
\format{
An object of class \code{list} of length 4.
}
\usage{
flexgsea_weighted_ks
}
\description{
The weighted KS-like statistic as described by Subramanian et al (2005).
}
\references{
Subramanian, A. et al. (2005) Gene Set Enrichment Analysis: A
  Knowledge-Based Approach for Interpreting Genome-Wide Expression
  Profiles. \emph{PNAS} \strong{102} (43): 15545-50.
  doi:10.1073/pnas.0506580102.
}
\seealso{
Other gene set enrichment functions: 
\code{\link{flexgsea_maxmean}},
\code{\link{flexgsea_mean}}
}
\concept{gene set enrichment functions}
\keyword{datasets}
