% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsea.R
\name{ggsea_calc_sig}
\alias{ggsea_calc_sig}
\title{Calculate significance of gene set enrichments from permutations.}
\usage{
ggsea(x, y, gene.sets, sig.fun=ggsea_calc_sig)
}
\description{
Calculates significance like Subramanian et al.. Significance is calculated
separately for positive and ES scores. A normalized enrichment score (NES)
is calculated. The NES for all gene sets are combined to get more precision
with less permutations.
}
\references{
Subramanian, A. et al. (2005) Gene Set Enrichment Analysis: A
  Knowledge-Based Approach for Interpreting Genome-Wide Expression
  Profiles. \emph{PNAS} \strong{102} (43): 15545–50.
  doi:10.1073/pnas.0506580102.
}
\seealso{
Other functions for significance calculation: \code{\link{ggsea_calc_sig_simple}}
}
