% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsea.R
\name{ggsea_calc_sig_simple}
\alias{ggsea_calc_sig_simple}
\title{Calculate significance of gene set enrichments from permutations.}
\usage{
ggsea(x, y, gene.sets, sig.fun=ggsea_calc_sig_simple)
}
\description{
Calculates significance by the rank of ES score in the permuted values.
Significance is calculated separately per gene set.
}
\seealso{
Other functions for significance calculation: \code{\link{ggsea_calc_sig}}
}
