% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsea_limma.R
\name{ggsea_limma}
\alias{ggsea_limma}
\title{Score genes with limma with a design matrix.}
\usage{
ggsea_limma_dm
}
\description{
Scores genes by their coefficients in from limma. 
The argument \code{y} should be a design matrix, as produced by
\code{model.matrix}. 
For RNA-seq data, use the
\code{EList} object produced by the \code{\link[limma]{voom}} function.
Do not call directly, but give as the \option{gene.score.fn}  argument to
\code{\link{ggsea}}.
}
\seealso{
Other gene scoring functions: \code{\link{ggsea_lm}},
  \code{\link{ggsea_s2n}}
}
