% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsea.R
\name{ggsea_lm}
\alias{ggsea_lm}
\title{Score genes in a linear regression model.}
\usage{
ggsea_lm
}
\description{
Scores genes by their coefficients in a linear regression
model of \option{y} onto \option{x}. When there is one response variable
this is equivalent to Pearson correlation.
Do not call directly, but give as the \option{gene.score.fn}  argument to
\code{\link{ggsea}}.
}
\seealso{
Other gene scoring functions: \code{\link{ggsea_limma}},
  \code{\link{ggsea_s2n}}
}
