% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsea.R
\docType{data}
\name{ggsea_maxmean}
\alias{ggsea_maxmean}
\title{The maxmean statistic for calculating gene set enrichement scores.}
\format{An object of class \code{list} of length 3.}
\usage{
ggsea(x, y, gene.sets, es.fn=ggsea_maxmean)
}
\description{
The maxmean statistic for calculating gene set enrichement scores.
}
\seealso{
Other gene set enrichment functions: \code{\link{ggsea_mean}},
  \code{\link{ggsea_weighted_ks}}
}
\keyword{datasets}
