% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsea.R
\docType{data}
\name{ggsea_mean}
\alias{ggsea_mean}
\title{The mean statistic for calculating gene set enrichement scores.}
\format{An object of class \code{list} of length 3.}
\usage{
ggsea_mean
}
\description{
The mean statistic for calculating gene set enrichement scores.
}
\seealso{
Other gene set enrichment functions: \code{\link{ggsea_maxmean}},
  \code{\link{ggsea_weighted_ks}}
}
\keyword{datasets}
