% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsea.R
\name{ggsea_s2n}
\alias{ggsea_s2n}
\title{Score genes by there signal to noise ratio.}
\usage{
ggsea_s2n
}
\description{
Scores genes by the signal to noise ration (s2n) a. Requires \option{y}
to have two classes.
Do not call directly, but give as the \option{gene.score.fn}  argument to
\code{\link{ggsea}}.
}
\seealso{
Other gene scoring functions: \code{\link{ggsea_lm}}
}
