% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexgsea.R
\name{read_gmt}
\alias{read_gmt}
\title{Read gene sets in gmt format.}
\usage{
read_gmt(file, progress = interactive())
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data. Passed to
the \pkg{readr} package so also supports some compression and urls.}

\item{progress}{Display a progress bar? Passed to the \pkg{readr} package.}
}
\description{
Read gene sets in gmt format.
}
